pyproject.toml
[project]
name = "llmprompt-nexus"
version = "0.2.0"
description = "A unified framework for interacting with various LLM providers through a consistent interface with batch processing, templating, and rate limiting."
authors = [
    {name = "Ernesto Estevanell",email = "ernestoestevanellv@gmail.com"}
]
license = {text = "CC-BY-NC-4.0"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp (==3.9.3)",
    "pandas (==1.5.3)",
    "numpy (==1.26.4)",
    "tqdm (==4.66.2)",
    "python-dotenv (==1.0.0)",
    "pyyaml (==6.0.1)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
mypy = "^1.15.0"

